try launching scenarios from a command
