+++ test/react-native/features/fixtures/performance-scenarios/ConfigFileReader.js
