+++ test/react-native/features/fixtures/performance-scenarios/index.js
@@ -1,0 +1,1 @@
+export { launchScenario } from './ScenarioLauncher'
