+++ test/react-native/features/fixtures/performance-scenarios/package.json
