+++ test/react-native/features/fixtures/performance-scenarios/scenarios/Scenarios.js
@@ -1,0 +1,1 @@
+export { TestScenario } from './TestScenario'
