--- test/react-native/features/fixtures/scenarios/index.js
@@ -1,1 +1,0 @@
-export { Scenario } from './Scenario'
